Make it possible to ignore SSL errors from network communications
It seems  that the OpenSSL version  on some systems is  not capable of
correctly handling  intermediate certificates  for some  sites.  These
sites include SourceForge which we  use for serving sample files. This
change adds  a global handler for  QNAM SSL errors with  the option to
ignore them at the users discretion.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@7359 ab8295b8-cf94-4d9e-aec4-7959e3be5d79