Read mode before attempting to change it when using DX Lab Suite Commander
This change is  specifically to work around  some surprising behaviour
in the Elecraft K3[S] which suffers  audio drop outs when null changes
that might effect the internal DSP are processed.  This manifests when
a blind mode  change is sent through Commander. Here  we read the mode
and  only subsequently  command a  mode change  if it  is not  what we
expect.  Blind  mode changes were  preferred with Commander  until now
since state queries through Commander are occasionally unreliable.

This issue  has been reported to  Elecraft so may not  be permanent if
they can and will address it in the rig's firmware.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@7565 ab8295b8-cf94-4d9e-aec4-7959e3be5d79