Smarter Tx sequencing that allows for CAT processing latency
When using  split operating and/or rig  mode setting there is  a delay
before PTT  is asserted  while the CAT  commands are  processed. There
will even be  a small delay if  CAT is only used for  PTT. This change
waits for these  commands to be completed before starting  the user Tx
delay specified in the settings.

There is  a notable exception to  this in that DX  Lab Suite Commander
does not give us a way of knowing when CAT commands have actually been
completed.   This combined  with  the way  that  Commander spaces  CAT
commands, rather  than being paced  by the rig's CAT  responses, means
that delays of more than 1 second are likely. As most rigs suffer some
sort of ALC overshoot when PTT is asserted with audio already present,
it is strongly recommended not to use  the DX Lab Suite Commander as a
rig control proxy if your rig may  suffer ALC overshoot and you use an
external PA. Please note that using a typical RF power meter to detect
ALC  overshoots is  unwise  as  they are  rarely  fast  enough. It  is
probably wise to assume that all rigs suffer ALC overshoot since it is
very difficult to design an ALC loop that avoids it.

This change  also increases  the range of  the "Settings->Advanced->Tx
Delay" spin box to 0s up to 2s.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@7756 ab8295b8-cf94-4d9e-aec4-7959e3be5d79