Fix an issue with VFO tuning while running Doppler correction
Holding down  the SHIFT  key while  tuning the  rig should  update the
nominal  sked frequency,  not holding  done the  SHIFT key  should not
update the  sked frequency. This  is not  yet perfect and  sometimes a
change to the nominal sked frequency can  get through but it is a rare
as yet unfound race condition.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@7992 ab8295b8-cf94-4d9e-aec4-7959e3be5d79