Restrict automatic recognition of CQ nnn calls to exactly nnn of 010 through 999 inclusive
The  values of  CQ 000  through  CQ 009  are reserved  for calling  by
numbers  to reduce  a pile  up e.g.  "CQ 2  P5/G4WJS" indicating  only
callers with  a 2 in their  callsign should call. Numeric  balues with
less  than 3  digits or  more  3 digits  are not  allowed in  standard
messages.

This is in line with the message source encoding protocol.

Thanks to Mike, W9MDB, for the cotribution this change was based on.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@8572 ab8295b8-cf94-4d9e-aec4-7959e3be5d79