Tentative reversal of recent change to set hound Tx frequency very late
Changing the Hound  Tx audio offset just before starting  the Tx audio
stream is way too  late to have the Tx message  printed in the decodes
window with the correct frequency shown. If we intend to change the Tx
offset very  late like this  we are going to  be in trouble  with rigs
that  cannot do  CAT  while  transmitting since  PTT  will already  be
asserted.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@8627 ab8295b8-cf94-4d9e-aec4-7959e3be5d79