Improve handling of country/prefix information appended to decoded CQ messages
The  use  of  QTextEdit::toPlainText() converts  non-break  spaces  to
normal spaces so  was breaking the use of that  character to delineate
appended info. Since using  QTextEdit::toPlainText() is an unnecessary
operation as  the decoded text  can be more amenably  accessed through
the QDocument  and QTextBlock interface without  any translation, this
is an improvement which should  have some performance impact when many
decodes are stored in the decodes windows.