Allow for tx periods shorter than the audio buffer size
Audio will not start until at least one buffer full is achieved and as
we use a large target latency of 2s to minimize CPU usage and glitches
we must pad with silence when  the QAudioOutput pulls buffers from the
Modulator i/o device. This is  all necessary with pulseaudio using the
underlying o/s ALSA device, i.e. on Linux.