Better ADIF conformance for ARRL Field Day mode
This change may break N1MM Logger+ integration, notably the CLASS ADIF
field is  populated which may not  be recognized by N1MM  Logger+, nor
interfaces to it.

One exception to  ADIF conformance is that the ARRL_SECT  field may be
populated with  the value DX  despite it  not being a  valid ARRL_SECT
enumeration value. This is done for consistency with N1MM Logger+ ADIF
exports.