Environment variable to add extra data to azel.dat
The  new  environment  variable   WSJT_AZEL_EXTRA_LINES  when  set  to
positive integer value  adds up to that number of  extra data lines to
the azel.dat  file. Currently only  one extra line is  supported which
contains the following data: Dpol, MNR, Dgrd. Where the values are the
same as  displayed in the  astronomical data window. I.e.  EME spatial
polarization offset in degrees, maximum non-reciprocity of the PATH in
dB, and the estimated signal  degradation relative to best possible in
dB.